cosas de json y tal para guardar los datos

+++ app/src/main/java/com/example/myapplication/Tarjeta.kt
@@ -4,12 +4,12 @@ class Tarjeta(nombre: String) {
     var nombre: String = nombre
     var descripcion: String = ""
 
-    var altitud: Int = 0
-    var latitud: Int = 0
+    var altitud: Double = 0.0
+    var latitud: Double = 0.0
 
     // Constructor
 
-    constructor(nombre: String, descripcion: String, altitud: Int, latitud: Int) : this(nombre){
+    constructor(nombre: String, descripcion: String, altitud: Double, latitud: Double) : this(nombre){
         this.descripcion = descripcion
         this.altitud = altitud
         this.latitud = latitud
